Adding windows scripts

+++ src/scripts/devspace/clean.ps1
@@ -1,5 +1,7 @@
-containerid=$(docker ps -q -f name=kindest-argo-cd_devcontainer-kindest)
-docker rm -f $containerid -f
+$containerid=$(docker ps -q -f name=kindest-argo-cd_devcontainer-kindest)
+if ($containerid) {
+    docker rm -f $containerid -f
+}
 docker container prune -f
 docker image prune -a -f
 docker network prune -f
